+++ src/test/java/com/ajlopez/blockchain/test/utils/FactoryHelper.java
@@ -228,7 +228,7 @@ public class FactoryHelper {
     }
 
     public static NodeProcessor createNodeProcessor(BlockChain blockChain) {
-        return new NodeProcessor(new NetworkConfiguration(42), createRandomPeer(), blockChain, new TrieStore(new HashMapStore()), createRandomAddress());
+        return new NodeProcessor(new NetworkConfiguration((short)42), createRandomPeer(), blockChain, new TrieStore(new HashMapStore()), createRandomAddress());
     }
 
     public static List<Block> createBlocks(int nblocks) {
